api/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic # FastAPI dependencies